news/requirements.txt
attrs==23.2.0
Automat==24.8.1
Babel==2.10.3
bcc==0.29.1
blinker==1.7.0
Brlapi==0.8.5
Brotli==1.1.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloud-init==24.3.1
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
cryptography==41.0.7
cssselect==1.2.0
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
defusedxml==0.7.1
distro==1.9.0
distro-info==1.7+build1
dnspython==2.7.0
feedparser==6.0.11
filelock==3.16.1
fsspec==2024.10.0
gyp==0.1
httplib2==0.20.4
huggingface-hub==0.26.2
hyperlink==21.0.0
idna==3.6
incremental==24.7.2
iniconfig==2.0.0
itemadapter==0.9.0
itemloaders==1.3.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
netaddr==0.8.0
netifaces==0.11.0
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
olefile==0.46
packaging==24.1
parsel==1.9.1
pexpect==4.9.0
pillow==10.2.0
pluggy==1.5.0
Protego==0.3.1
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.25.1
pycups==2.0.1
PyDispatcher==2.0.7
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
pymongo==4.10.1
pyOpenSSL==24.2.1
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.0.1
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
queuelib==1.7.0
regex==2024.11.6
requests==2.31.0
requests-file==2.1.0
rich==13.7.1
safetensors==0.4.5
screen-resolution-extra==0.0.0
service-identity==24.2.0
setuptools==68.1.2
sgmllib3k==1.0.0
six==1.16.0
sympy==1.13.1
systemd-python==235
tldextract==5.1.2
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.0
transformers==4.46.2
triton==3.1.0
Twisted==24.10.0
typing_extensions==4.10.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
w3lib==2.2.1
wadllib==1.3.6
wheel==0.42.0
xdg==5
xkit==0.0.0
zope.interface==7.1.1
